4222016: Added (older) TextRender Test for PS
This was used to see what settings in the application manifest were
necessary for proper rendering on high-DPI screens, since I was doing a
lot of development at that time on screens that exceeded the industry
standard, and it became frustrating to look at with improper text
rendering (especially) in Snowshoes. This form is really, really simple,
but I'm sure I'll revisit it.