7/13/2016: It loads processes...but can you see them?
The way that this grid stacks is a bit strange...and everything that I
was going crazy about with Visual Studio and WPF I now dread. Get me out
of here. It's a glossy, well-aligned, beautiful realm, but one that
breaks or introduces way too many things that cause ambiguity errors.

I'm sure I'll take all that back, though, by this time tomorrow.