7162016: Worked through example on Win32 API
In the past, I've only used the Win32 API when I absolutely needed to,
properly handling windows, attributed processes and handles, etc. Today,
I realized that if I'm serious about proceeding with C++ and upcoming
versions of Snowshoes written in it, I should probably get familiar with
this again on  a larger scale.

I also forgot to name the project for some reason...hopefully that
doesn't happen ever again.