Set up for Process Playtime
I'm going to try and dig deeper in an effort to use less resources for
calling process changes, particularly for Snowshoes purposes (but for
any that might come up). WPF form, handled in C#, I just need to see
something quick that doesn't run on a console (need expensive draws to
happen).